use ENV=development instead of isDevelopmentMode=enabled
at the time of commit, both ENV and isDevelopmentMode are used interchangeably across the
codebase to check if the application is in development mode. This change ensures that
we are only checking one environment variable as a single source of truth to check if the
application is in development mode or not.